# Rework Mapper as type class, introduce ToTwitterFuture type class