--- a/core/src/main/scala/io/finch/syntax/EndpointMapper.scala
+++ b/core/src/main/scala/io/finch/syntax/EndpointMapper.scala
@@ -8,7 +8,7 @@ class EndpointMapper[A](m: Method, e: Endpoint[A]) extends Endpoint[A] { self =>
   /**
    * Maps this endpoint to either `A => Output[B]` or `A => Future[Output[B]]`.
    */
-  final def apply(mapper: Mapper[A]): Endpoint[mapper.Out] = mapper(self)
+  final def apply[F](f: F)(implicit mapper: Mapper[F, A]): Endpoint[mapper.Out] = mapper(f, self)
 
   final def apply(input: Input): Endpoint.Result[A] =
     if (input.request.method == m) e(input)