--- a/core/src/main/scala/io/finch/syntax/EndpointMapper.scala
+++ b/core/src/main/scala/io/finch/syntax/EndpointMapper.scala
@@ -2,6 +2,7 @@ package io.finch.syntax
 
 import com.twitter.finagle.http.Method
 import io.finch._
+import shapeless.Lazy
 
 class EndpointMapper[A](m: Method, e: Endpoint[A]) extends Endpoint[A] { self =>