--- a/core/src/main/scala/io/finch/syntax/Mapper.scala
+++ b/core/src/main/scala/io/finch/syntax/Mapper.scala
@@ -1,7 +1,7 @@
 package io.finch.syntax
 
+import cats.Id
 import com.twitter.finagle.http.Response
-import com.twitter.util.Future
 import io.finch.{Endpoint, Output}
 import shapeless.HNil
 import shapeless.ops.function.FnToProduct
@@ -13,43 +13,48 @@ import shapeless.ops.function.FnToProduct
  * @groupname HighPriorityMapper High priority mapper conversions
  * @groupprio HighPriorityMapper 1
  */
-trait Mapper[A] {
+trait Mapper[F, A] {
   type Out
 
-  def apply(e: Endpoint[A]): Endpoint[Out]
+  def apply(f: F, e: Endpoint[A]): Endpoint[Out]
 }
 
 private[finch] trait LowPriorityMapperConversions {
-  type Aux[A, B] = Mapper[A] { type Out = B }
+  type Aux[F, A, B] = Mapper[F, A] { type Out = B }
 
-  def instance[A, B](f: Endpoint[A] => Endpoint[B]): Mapper.Aux[A, B] = new Mapper[A] {
+  def instance[F, A, B](fn: (Endpoint[A], F) => Endpoint[B]): Mapper.Aux[F, A, B] = new Mapper[F, A] {
     type Out = B
-    def apply(e: Endpoint[A]): Endpoint[B] = f(e)
+
+    override def apply(f: F, e: Endpoint[A]): Endpoint[B] = fn(e, f)
   }
 
   /**
     * @group LowPriorityMapper
     */
-  implicit def mapperFromOutputFunction[A, B](f: A => Output[B]): Mapper.Aux[A, B] =
-    instance(_.mapOutput(f))
+  implicit def mapperFromOutputFunction[A, B]: Mapper.Aux[A => Output[B], A, B] =
+    instance(_ mapOutput _)
 
   /**
    * @group LowPriorityMapper
    */
-  implicit def mapperFromResponseFunction[A](f: A => Response): Mapper.Aux[A, Response] =
-    instance(_.mapOutput(f.andThen(r => Output.payload(r, r.status))))
+  implicit def mapperFromResponseFunction[A]: Mapper.Aux[A => Response, A, Response] =
+    instance(_ mapOutput _.andThen(r => Output.payload(r, r.status)))
 
   /**
     * @group LowPriorityMapper
     */
-  implicit def mapperFromFutureOutputFunction[A, B](f: A => Future[Output[B]]): Mapper.Aux[A, B] =
-    instance(_.mapOutputAsync(f))
+  implicit def mapperFromHOutputFunction[F[_], A, B](implicit
+                                                     ttf: ToTwitterFuture[F]
+                                                    ): Mapper.Aux[A => F[Output[B]], A, B] =
+    instance(_ mapOutputAsync _.andThen(ttf.apply))
 
   /**
    * @group LowPriorityMapper
    */
-  implicit def mapperFromFutureResponseFunction[A](f: A => Future[Response]): Mapper.Aux[A, Response] =
-    instance(_.mapOutputAsync(f.andThen(fr => fr.map(r => Output.payload(r, r.status)))))
+  implicit def mapperFromHResponseFunction[F[_], A](implicit
+    ttf: ToTwitterFuture[F]
+  ): Mapper.Aux[A => F[Response], A, Response] =
+    instance(_ mapOutputAsync _.andThen(ttf.apply).andThen(fr => fr.map(r => Output.payload(r, r.status))))
 }
 
 private[finch] trait HighPriorityMapperConversions extends LowPriorityMapperConversions {
@@ -57,19 +62,19 @@ private[finch] trait HighPriorityMapperConversions extends LowPriorityMapperConv
   /**
     * @group HighPriorityMapper
     */
-  implicit def mapperFromOutputHFunction[A, B, F, OB](f: F)(implicit
+  implicit def mapperFromOutputHFunction[A, B, F, OB](implicit
     ftp: FnToProduct.Aux[F, A => OB],
-    ev: OB <:< Output[B]
-  ): Mapper.Aux[A, B] = instance(_.mapOutput(value => ev(ftp(f)(value))))
+    ev: OB <:< Id[Output[B]]
+  ): Mapper.Aux[F, A, B] = instance((e, f) => e.mapOutput[B](value => ev(ftp(f)(value))))
 
 
   /**
    * @group HighPriorityMapper
    */
-  implicit def mapperFromResponseHFunction[A, F, R](f: F)(implicit
+  implicit def mapperFromResponseHFunction[A, F, R](implicit
     ftp: FnToProduct.Aux[F, A => R],
     ev: R <:< Response
-  ): Mapper.Aux[A, Response] = instance(_.mapOutput { value =>
+  ): Mapper.Aux[F, A, Response] = instance((e, f) => e.mapOutput { value =>
     val r = ev(ftp(f)(value))
     Output.payload(r, r.status)
   })
@@ -77,39 +82,45 @@ private[finch] trait HighPriorityMapperConversions extends LowPriorityMapperConv
   /**
    * @group HighPriorityMapper
    */
-  implicit def mapperFromOutputValue[A](o: => Output[A]): Mapper.Aux[HNil, A] =
-    instance(_.mapOutput(_ => o))
+  implicit def mapperFromOutputValue[A]: Mapper.Aux[Output[A], HNil, A] =
+    instance((e, o) => e.mapOutput(_ => o))
 
   /**
    * @group HighPriorityMapper
    */
-  implicit def mapperFromResponseValue(r: => Response): Mapper.Aux[HNil, Response] =
-    instance(_.mapOutput(_ => Output.payload(r, r.status)))
+  implicit def mapperFromResponseValue: Mapper.Aux[Response, HNil, Response] =
+    instance((e, r) => e.mapOutput(_ => Output.payload(r, r.status)))
 
   /**
     * @group HighPriorityMapper
     */
-  implicit def mapperFromFutureOutputValue[A](o: => Future[Output[A]]): Mapper.Aux[HNil, A] =
-    instance(_.mapOutputAsync(_ => o))
+  implicit def mapperFromHOutputValue[F[_], A](implicit
+    ttf: ToTwitterFuture[F]
+  ): Mapper.Aux[F[Output[A]], HNil, A] =
+    instance((e, f) => e.mapOutputAsync(_ => ttf(f)))
 
   /**
    * @group HighPriorityMapper
    */
-  implicit def mapperFromFutureResponseValue(fr: => Future[Response]): Mapper.Aux[HNil, Response] =
-    instance(_.mapOutputAsync(_ => fr.map(r => Output.payload(r, r.status))))
+  implicit def mapperFromHResponseValue[F[_]](implicit
+    ttf: ToTwitterFuture[F]
+  ): Mapper.Aux[F[Response], HNil, Response] =
+    instance((e, f) => e.mapOutputAsync(_ => ttf(f).map(r => Output.payload(r, r.status))))
 }
 
 object Mapper extends HighPriorityMapperConversions {
-  implicit def mapperFromFutureOutputHFunction[A, B, F, FOB](f: F)(implicit
-    ftp: FnToProduct.Aux[F, A => FOB],
-    ev: FOB <:< Future[Output[B]]
-  ): Mapper.Aux[A, B] = instance(_.mapOutputAsync(value => ev(ftp(f)(value))))
+  implicit def mapperFromHOutputHFunction[F[_], A, B, FN, FOB](implicit
+    ftp: FnToProduct.Aux[FN, A => FOB],
+    ev: FOB <:< F[Output[B]],
+    ttf: ToTwitterFuture[F]
+  ): Mapper.Aux[FN, A, B] = instance((e, f) => e.mapOutputAsync(value => ttf(ev(ftp(f)(value)))))
 
-  implicit def mapperFromFutureResponseHFunction[A, F, FR](f: F)(implicit
-    ftp: FnToProduct.Aux[F, A => FR],
-    ev: FR <:< Future[Response]
-  ): Mapper.Aux[A, Response] = instance(_.mapOutputAsync { value =>
-    val fr = ev(ftp(f)(value))
+  implicit def mapperFromHResponseHFunction[F[_], A, FN, FR](f: FN)(implicit
+    ftp: FnToProduct.Aux[FN, A => FR],
+    ev: FR <:< F[Response],
+    ttf: ToTwitterFuture[F]
+  ): Mapper.Aux[FN, A, Response] = instance((e, f) => e.mapOutputAsync { value =>
+    val fr = ttf(ev(ftp(f)(value)))
     fr.map(r => Output.payload(r, r.status))
   })
 }