--- a/core/src/main/scala/io/finch/syntax/ToTwitterFuture.scala
+++ b/core/src/main/scala/io/finch/syntax/ToTwitterFuture.scala
@@ -1,0 +1,17 @@
+package io.finch.syntax
+
+import com.twitter.util.Future
+
+trait ToTwitterFuture[F[_]] {
+
+  def apply[A](f: F[A]): Future[A]
+
+}
+
+object ToTwitterFuture {
+
+  implicit val identity: ToTwitterFuture[Future] = new ToTwitterFuture[Future] {
+    override def apply[A](f: Future[A]): Future[A] = f
+  }
+
+}