--- a/core/src/main/scala/io/finch/syntax/ToTwitterFuture.scala
+++ b/core/src/main/scala/io/finch/syntax/ToTwitterFuture.scala
@@ -2,6 +2,9 @@ package io.finch.syntax
 
 import com.twitter.util.Future
 
+/**
+  * Type class for conversion of some HKT (i.e. scala.concurrent.Future) to twitter future
+  */
 trait ToTwitterFuture[F[_]] {
 
   def apply[A](f: F[A]): Future[A]
@@ -11,7 +14,7 @@ trait ToTwitterFuture[F[_]] {
 object ToTwitterFuture {
 
   implicit val identity: ToTwitterFuture[Future] = new ToTwitterFuture[Future] {
-    override def apply[A](f: Future[A]): Future[A] = f
+    def apply[A](f: Future[A]): Future[A] = f
   }
 
 }