--- a/core/src/main/scala/io/finch/syntax/scala/package.scala
+++ b/core/src/main/scala/io/finch/syntax/scala/package.scala
@@ -1,0 +1,3 @@
+package io.finch.syntax
+
+package object scala extends ScalaToTwitterFuture