Mapper now is not a magnet pattern, but a type class. Signature of `EndpointMapper.apply` has changed a little, but it's backward compatible. Strictly speaking, from API point of view nothing was changed, except that IDE should stop whining about "wrong" arguments: https://github.com/finagle/finch/issues/631

At the same time, added `ToTwitterFuture` type class to support different async mechanisms. There is an identity for twitter future and scala futures support in different package. Inspired by https://github.com/finagle/finch/pull/815

Potentially, this PR could open a way to support monix and scalaz tasks as parts of different subprojects: https://github.com/finagle/finch/pull/815#issuecomment-318662893

It's a working code, but without specs. If it's valid from "ideological" PoV I could write proper tests for Mapper.