@vkostyukov 2017-08-07
@ImLiar Thanks for working on this! Have you checked IntelliJ to see if it complains about endpoints now?