@vkostyukov 2017-08-07
That's really nice! Let me pull this locally and play with it a bit.