@vkostyukov 2017-09-04
I think this would also resolve #383.