@ImLiar 2017-09-04
Let me add tests here first anyway