@ImLiar 2017-09-08
Hi

I'm on shorty vacation for few days. Going to finish on Sunday
On Fri, 8 Sep 2017 at 19.17, Vladimir Kostyukov <notifications@github.com>
wrote:

> Hey @ImLiar <https://github.com/imliar>! How is this going? Do you need
> any help?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/finagle/finch/pull/816#issuecomment-328148699>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/ACNbQVhVx41mzgSYJwak9glCqMWMcMIDks5sgWivgaJpZM4Otx_g>
> .
>