@ImLiar 2017-09-11
@vkostyukov @rpless 

Benchmark results for this PR and master:

```
PR:
Benchmark                               Mode  Cnt    Score     Error  Units
ApplyBenchmark.function1FutureOutput    avgt   10   61.254 ±  10.093  ns/op
ApplyBenchmark.function1FutureResponse  avgt   10  110.490 ±  73.496  ns/op
ApplyBenchmark.function1Output          avgt   10  116.603 ± 107.588  ns/op
ApplyBenchmark.function1Response        avgt   10   71.227 ±  23.770  ns/op
ApplyBenchmark.functionNFutureOutput    avgt   10  142.614 ±  89.723  ns/op
ApplyBenchmark.functionNFutureResponse  avgt   10  116.035 ±  88.242  ns/op
ApplyBenchmark.functionNOutput          avgt   10   82.280 ±   9.138  ns/op
ApplyBenchmark.functionNResponse        avgt   10  135.343 ±  19.729  ns/op
ApplyBenchmark.futureOutput             avgt   10    7.761 ±   0.613  ns/op
ApplyBenchmark.futureResponse           avgt   10    7.410 ±   0.192  ns/op
ApplyBenchmark.output                   avgt   10    7.609 ±   0.606  ns/op
ApplyBenchmark.response                 avgt   10    7.403 ±   0.229  ns/op

Master:
Benchmark                               Mode  Cnt    Score    Error  Units
ApplyBenchmark.function1FutureOutput    avgt   10   66.755 ± 12.514  ns/op
ApplyBenchmark.function1FutureResponse  avgt   10   68.049 ± 11.998  ns/op
ApplyBenchmark.function1Output          avgt   10   66.912 ± 14.796  ns/op
ApplyBenchmark.function1Response        avgt   10   95.094 ± 56.855  ns/op
ApplyBenchmark.functionNFutureOutput    avgt   10  124.046 ± 33.424  ns/op
ApplyBenchmark.functionNFutureResponse  avgt   10   76.451 ± 13.739  ns/op
ApplyBenchmark.functionNOutput          avgt   10   82.871 ±  8.452  ns/op
ApplyBenchmark.functionNResponse        avgt   10  148.733 ± 30.853  ns/op
ApplyBenchmark.futureOutput             avgt   10    7.430 ±  0.117  ns/op
ApplyBenchmark.futureResponse           avgt   10    7.338 ±  0.192  ns/op
ApplyBenchmark.output                   avgt   10    7.738 ±  0.749  ns/op
ApplyBenchmark.response                 avgt   10    7.849 ±  0.728  ns/op

PR:
Benchmark                                                             Mode  Cnt     Score     Error   Units
ApplyBenchmark.function1FutureOutput:·gc.alloc.rate.norm              avgt   10   188.000 ±  19.124    B/op
ApplyBenchmark.function1FutureResponse:·gc.alloc.rate.norm            avgt   10   188.000 ±  19.124    B/op
ApplyBenchmark.function1Output:·gc.alloc.rate.norm                    avgt   10   200.000 ±   0.001    B/op
ApplyBenchmark.function1Response:·gc.alloc.rate.norm                  avgt   10   176.000 ±   0.001    B/op
ApplyBenchmark.functionNFutureOutput:·gc.alloc.rate.norm              avgt   10   168.000 ±   0.001    B/op
ApplyBenchmark.functionNFutureResponse:·gc.alloc.rate.norm            avgt   10   168.000 ±   0.001    B/op
ApplyBenchmark.functionNOutput:·gc.alloc.rate.norm                    avgt   10   168.000 ±   0.001    B/op
ApplyBenchmark.functionNResponse:·gc.alloc.rate.norm                  avgt   10   420.000 ±  19.124    B/op
ApplyBenchmark.futureOutput:·gc.alloc.rate.norm                       avgt   10    ≈ 10⁻⁵              B/op
ApplyBenchmark.futureResponse:·gc.alloc.rate.norm                     avgt   10    ≈ 10⁻⁵              B/op
ApplyBenchmark.output:·gc.alloc.rate.norm                             avgt   10    ≈ 10⁻⁵              B/op
ApplyBenchmark.response:·gc.alloc.rate.norm                           avgt   10    ≈ 10⁻⁵              B/op

Master:
Benchmark                                                             Mode  Cnt     Score     Error   Units
ApplyBenchmark.function1FutureOutput:·gc.alloc.rate.norm              avgt   10   188.000 ±  19.124    B/op
ApplyBenchmark.function1FutureResponse:·gc.alloc.rate.norm            avgt   10   176.000 ±   0.001    B/op
ApplyBenchmark.function1Output:·gc.alloc.rate.norm                    avgt   10   188.000 ±  19.124    B/op
ApplyBenchmark.function1Response:·gc.alloc.rate.norm                  avgt   10   188.000 ±  19.124    B/op
ApplyBenchmark.functionNFutureOutput:·gc.alloc.rate.norm              avgt   10   168.000 ±   0.001    B/op
ApplyBenchmark.functionNFutureResponse:·gc.alloc.rate.norm            avgt   10   168.000 ±   0.001    B/op
ApplyBenchmark.functionNOutput:·gc.alloc.rate.norm                    avgt   10   168.000 ±   0.001    B/op
ApplyBenchmark.functionNResponse:·gc.alloc.rate.norm                  avgt   10   412.000 ±   6.375    B/op
ApplyBenchmark.futureOutput:·gc.alloc.rate.norm                       avgt   10    ≈ 10⁻⁵              B/op
ApplyBenchmark.futureResponse:·gc.alloc.rate.norm                     avgt   10    ≈ 10⁻⁵              B/op
ApplyBenchmark.output:·gc.alloc.rate.norm                             avgt   10    ≈ 10⁻⁵              B/op
ApplyBenchmark.response:·gc.alloc.rate.norm                           avgt   10    ≈ 10⁻⁵              B/op
```

So there is a quite a difference in overhead. By the way, while working with this PR, I realised that my latest IDEA highlighting works correctly with master for every case but gives an error for `Future[Output[A]]` case in PR... Damn tools

How about postpone decision about this one and move `ToTwitterFuture` to a separate PR? It would open a door for different async libraries meanwhile