@vkostyukov 2017-10-02
@ImLiar What's the plan on this? Do we want to continue working in that direction or not?