@ImLiar 2017-10-10
@vkostyukov hey

I think I could finish this one, but it won't help 100% with highlighting.